DC Setup
Did some basic preparation for switching to using the dynamic camera
class.
Created the DynamicCamera class and implemented some basic methods.
Added the Camera interface and the default generateQuad() method.
Improved greatly on the Rectangle class.
Added the Measurement abstract class.
Made it so that VisibleObjects have an area(Rectangle) by which they are
represented and bound its center to the object's location, (have not
bound the size changes yet)