Misc
Now able to compare stats between different tracker entities.
Time is now logged before the main run loop starts (no more hundrends of
ticks executing as fast as possible while the time variable tries to
catch up to the current time)
Fixed an exception that was occuring because some of the code was using
the standard library mouse listeners that cannot receive input that is
intercepted by the canvas.
Added Calendar