Edge rendering and ability fix
The should be none (or at least less) problems when trying to add
abilities to a hero.
Dynamic camera is now able to draw only edges of actors.
Rectangle now show their center and dimensions (as opposed to the
northWest point and dimensions) when its toString() is called.
Decreased the visibility of the abilityButtons list in GUI from public
to package to avoid API confusion.